Re-Organizing BLE into Multiple Files
